backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
sqlalchemy==2.0.23
pydantic==2.4.2
python-dotenv==1.0.0
alembic==1.12.1
pytest==7.4.3
httpx==0.25.1
python-multipart==0.0.6
psycopg2-binary==2.9.9
asyncpg==0.29.0 